+++ src/main/kotlin/icu/windea/pls/extension/diagram/DiagramElementManagerEx.kt
@@ -6,4 +6,3 @@ import javax.swing.*
 abstract class DiagramElementManagerEx<T> : AbstractDiagramElementManager<T>() {
     open fun getItemComponent(nodeElement: T, nodeItem: Any?, builder: DiagramBuilder): JComponent? = null
 }
-