+++ src/main/kotlin/icu/windea/pls/extension/diagram/provider/ParadoxEventTreeDiagramProvider.kt
@@ -5,19 +5,19 @@ import com.intellij.diagram.presentation.*
 import com.intellij.openapi.application.*
 import com.intellij.openapi.progress.*
 import com.intellij.openapi.project.*
+import com.intellij.openapi.util.*
 import com.intellij.openapi.vfs.*
 import com.intellij.psi.*
 import com.intellij.ui.*
 import icu.windea.pls.*
+import icu.windea.pls.config.configGroup.*
+import icu.windea.pls.config.util.*
 import icu.windea.pls.core.*
-import icu.windea.pls.core.util.*
-import icu.windea.pls.cwt.psi.*
 import icu.windea.pls.extension.diagram.*
 import icu.windea.pls.extension.diagram.settings.*
 import icu.windea.pls.lang.*
 import icu.windea.pls.lang.util.*
 import icu.windea.pls.lang.util.renderers.*
-import icu.windea.pls.localisation.psi.*
 import icu.windea.pls.model.*
 import icu.windea.pls.model.constants.*
 import icu.windea.pls.script.psi.*
@@ -31,15 +31,32 @@ import javax.swing.*
  * * 支持任何通用的图表操作。（例如，导出为图片）
  */
 abstract class ParadoxEventTreeDiagramProvider(gameType: ParadoxGameType) : ParadoxDefinitionDiagramProvider(gameType) {
-    companion object {
-        val CAT_TYPE = DiagramCategory(PlsDiagramBundle.lazyMessage("eventTree.category.type"), PlsIcons.Nodes.Type, true, false)
-        val CAT_PROPERTIES = DiagramCategory(PlsDiagramBundle.lazyMessage("eventTree.category.properties"), PlsIcons.Nodes.Property, true, false)
-        val CAT_TITLE = DiagramCategory(PlsDiagramBundle.lazyMessage("eventTree.category.title"), PlsIcons.Nodes.Localisation, false, false)
-        val CAT_PICTURE = DiagramCategory(PlsDiagramBundle.lazyMessage("eventTree.category.picture"), PlsIcons.General.Image, false, false)
-        val CATEGORIES = arrayOf(CAT_TYPE, CAT_PROPERTIES, CAT_TITLE, CAT_PICTURE)
+    object Categories {
+        val Type = DiagramCategory(PlsDiagramBundle.lazyMessage("eventTree.category.type"), PlsIcons.Nodes.Type, true, false)
+        val Properties = DiagramCategory(PlsDiagramBundle.lazyMessage("eventTree.category.properties"), PlsIcons.Nodes.Property, true, false)
+        val Title = DiagramCategory(PlsDiagramBundle.lazyMessage("eventTree.category.title"), PlsIcons.Nodes.Localisation, false, false)
+        val Picture = DiagramCategory(PlsDiagramBundle.lazyMessage("eventTree.category.picture"), PlsIcons.General.Image, false, false)
 
-        val REL_INVOKE = object : DiagramRelationshipInfoAdapter("INVOKE", DiagramLineType.SOLID, PlsDiagramBundle.message("eventTree.rel.invoke")) {
+        val All = arrayOf(Type, Properties, Title, Picture)
+    }
+
+    object Relations {
+        val Invoke = object : DiagramRelationshipInfoAdapter("INVOKE", DiagramLineType.SOLID, PlsDiagramBundle.message("eventTree.rel.invoke")) {
             override fun getTargetArrow() = DELTA
+        }
+    }
+
+    object Items {
+        class Type(val definition: ParadoxScriptProperty)
+
+        class Property(val property: ParadoxScriptProperty)
+
+        class Title(val definition: ParadoxScriptProperty) {
+            val nameElement by lazy { ParadoxTechnologyManager.getLocalizedNameElement(definition) }
+        }
+
+        class Picture(val definition: ParadoxScriptProperty) {
+            val iconFile by lazy { ParadoxTechnologyManager.getIconFile(definition) }
         }
     }
 
@@ -51,30 +68,23 @@ abstract class ParadoxEventTreeDiagramProvider(gameType: ParadoxGameType) : Para
 
     abstract override fun createDataModel(project: Project, element: PsiElement?, file: VirtualFile?, model: DiagramPresentationModel): DataModel
 
-    override fun getAllContentCategories() = CATEGORIES
+    override fun getAllContentCategories() = Categories.All
 
     abstract override fun getDiagramSettings(project: Project): ParadoxEventTreeDiagramSettings<*>?
 
     class NodeContentManager : OrderedDiagramNodeContentManager() {
         override fun isInCategory(nodeElement: Any?, item: Any?, category: DiagramCategory, builder: DiagramBuilder?): Boolean {
-            return when {
-                item is CwtProperty -> category == CAT_TYPE
-                item is ParadoxScriptProperty -> {
-                    val definitionInfo = item.definitionInfo
-                    if (definitionInfo != null) {
-                        false
-                    } else {
-                        category == CAT_PROPERTIES
-                    }
-                }
-                item is ParadoxLocalisationProperty -> category == CAT_TITLE
-                item is PsiFile -> category == CAT_PICTURE
-                else -> false
+            return when (item) {
+                is Items.Type -> category == Categories.Type
+                is Items.Property -> category == Categories.Properties
+                is Items.Title -> category == Categories.Title
+                is Items.Picture -> category == Categories.Picture
+                else -> true
             }
         }
 
         override fun getContentCategories(): Array<DiagramCategory> {
-            return CATEGORIES
+            return Categories.All
         }
     }
 
@@ -89,77 +99,51 @@ abstract class ParadoxEventTreeDiagramProvider(gameType: ParadoxGameType) : Para
 
         override fun getElementTitle(element: PsiElement): String? {
             ProgressManager.checkCanceled()
-            return runReadAction { doGetElementTitle(element) }
-        }
-
-        private fun doGetElementTitle(element: PsiElement): String? {
             return when (element) {
-                is PsiDirectory -> element.name
-                is ParadoxScriptProperty -> ParadoxEventManager.getName(element)
+                is PsiDirectory -> runReadAction { element.name }
+                is ParadoxScriptProperty -> runReadAction { ParadoxEventManager.getName(element) }
                 else -> null
             }
         }
 
         override fun getNodeItems(nodeElement: PsiElement?, builder: DiagramBuilder): Array<Any> {
+            provider as ParadoxEventTreeDiagramProvider
             ProgressManager.checkCanceled()
-            return runReadAction { doGetNodeItems(nodeElement) }
-        }
-
-        private fun doGetNodeItems(nodeElement: PsiElement?): Array<Any> {
             return when (nodeElement) {
                 is ParadoxScriptProperty -> {
                     val result = mutableListOf<Any>()
-                    val typeElement = nodeElement.definitionInfo?.typeConfig?.pointer?.element //should not be null
-                    if (typeElement != null) result.add(typeElement)
-                    val properties = getProperties(nodeElement)
-                    result.addAll(properties)
-                    val localizedNameElement = ParadoxEventManager.getLocalizedNameElement(nodeElement)
-                    if (localizedNameElement != null) result.add(localizedNameElement)
-                    val icon = ParadoxEventManager.getIconFile(nodeElement)
-                    if (icon != null) result.add(icon)
+                    result += Items.Type(nodeElement)
+                    val properties = runReadAction { provider.getProperties(nodeElement) }
+                    properties.forEach { result += Items.Property(it) }
+                    result += Items.Title(nodeElement)
+                    result += Items.Picture(nodeElement)
                     result.toTypedArray()
                 }
                 else -> emptyArray()
             }
         }
 
-        private fun getProperties(nodeElement: ParadoxScriptProperty): Set<ParadoxScriptProperty> {
-            provider as ParadoxEventTreeDiagramProvider
-            val itemPropertyKeys = provider.getItemPropertyKeys()
-            val properties = sortedSetOf<ParadoxScriptProperty>(compareBy { itemPropertyKeys.indexOf(it.name.lowercase()) })
-            nodeElement.block?.properties(conditional = true, inline = true)?.forEach {
-                if (it.name.lowercase() in itemPropertyKeys) properties.add(it)
-            }
-            return properties
-        }
-
         override fun getItemComponent(nodeElement: PsiElement, nodeItem: Any?, builder: DiagramBuilder): JComponent? {
             ProgressManager.checkCanceled()
-            return runReadAction { doGetItemComponent(nodeElement, nodeItem) }
-        }
+            return when (nodeItem) {
+                is Items.Title -> runReadAction r@{
+                    //科技的名字
+                    val nameElement = nodeItem.nameElement ?: return@r null
+                    val result = ParadoxLocalisationTextUIRenderer().render(nameElement)
+                    result
+                }
+                is Items.Picture -> runReadAction r@{
+                    //科技的图标
+                    val iconFile = nodeItem.iconFile ?: return@r null
+                    val frameInfo = nodeElement.getUserData(PlsKeys.imageFrameInfo)
+                    val iconUrl = ParadoxImageManager.resolveUrlByFile(iconFile.virtualFile, iconFile.project, frameInfo)
 
-        private fun doGetItemComponent(nodeElement: PsiElement, nodeItem: Any?): JComponent? {
-            return when (nodeElement) {
-                is ParadoxScriptProperty -> {
-                    when {
-                        nodeItem is ParadoxLocalisationProperty -> {
-                            //事件标题
-                            ParadoxLocalisationTextUIRenderer().render(nodeItem)
-                        }
-                        nodeItem is PsiFile -> {
-                            //事件图片
-                            val frameInfo = nodeElement.getUserData(PlsKeys.imageFrameInfo)
-                            val iconUrl = ParadoxImageManager.resolveUrlByFile(nodeItem.virtualFile, nodeItem.project, frameInfo)
+                    //如果无法解析（包括对应文件不存在的情况）就直接跳过
+                    if (!ParadoxImageManager.canResolve(iconUrl)) return@r null
 
-                            //如果无法解析（包括对应文件不存在的情况）就直接跳过
-                            if (!ParadoxImageManager.canResolve(iconUrl)) return null
-
-                            val iconFileUrl = iconUrl.toFileUrl()
-                            val icon = iconFileUrl.toIconOrNull()
-                            icon?.toLabel()
-                        }
-                        else -> null
-                    }
+                    val iconFileUrl = iconUrl.toFileUrl()
+                    val icon = iconFileUrl.toIconOrNull()
+                    icon?.toLabel()
                 }
                 else -> null
             }
@@ -167,59 +151,32 @@ abstract class ParadoxEventTreeDiagramProvider(gameType: ParadoxGameType) : Para
 
         override fun getItemName(nodeElement: PsiElement?, nodeItem: Any?, builder: DiagramBuilder): SimpleColoredText? {
             ProgressManager.checkCanceled()
-            return runReadAction { doGetItemName(nodeElement, nodeItem) }
-        }
-
-        private fun doGetItemName(nodeElement: PsiElement?, nodeItem: Any?): SimpleColoredText? {
-            return when (nodeElement) {
-                is ParadoxScriptProperty -> {
-                    when {
-                        nodeItem is CwtProperty -> {
-                            val typesText = nodeElement.definitionInfo?.typesText ?: return null
-                            val result = SimpleColoredText(typesText, DEFAULT_TEXT_ATTR)
-                            result
-                        }
-                        nodeItem is ParadoxScriptProperty -> {
-                            val definitionInfo = nodeItem.definitionInfo
-                            if (definitionInfo != null) return null
-                            val rendered = ParadoxScriptTextRenderer(renderInBlock = true).render(nodeItem)
-                            val result = SimpleColoredText(rendered, DEFAULT_TEXT_ATTR)
-                            val propertyValue = nodeItem.propertyValue
-                            if (propertyValue is ParadoxScriptScriptedVariableReference) {
-                                val sv = propertyValue.text
-                                result.append(" by $sv", SimpleTextAttributes.GRAYED_ATTRIBUTES)
-                            }
-                            result
-                        }
-                        else -> null
+            return when (nodeItem) {
+                is Items.Type -> runReadAction r@{
+                    val typesText = nodeItem.definition.definitionInfo?.typesText ?: return@r null
+                    val result = SimpleColoredText(typesText, DEFAULT_TEXT_ATTR)
+                    result
+                }
+                is Items.Property -> runReadAction {
+                    val property = nodeItem.property
+                    val rendered = ParadoxScriptTextRenderer(renderInBlock = true).render(property)
+                    val result = SimpleColoredText(rendered, DEFAULT_TEXT_ATTR)
+                    val propertyValue = property.propertyValue
+                    if (propertyValue is ParadoxScriptScriptedVariableReference) {
+                        val sv = propertyValue.text
+                        result.append(" by $sv", SimpleTextAttributes.GRAYED_ATTRIBUTES)
                     }
+                    result
                 }
                 else -> null
             }
         }
 
-        override fun getItemType(nodeElement: PsiElement?, nodeItem: Any?, builder: DiagramBuilder?): SimpleColoredText? {
-            return null
-        }
-
         override fun getItemIcon(nodeElement: PsiElement?, nodeItem: Any?, builder: DiagramBuilder?): Icon? {
             ProgressManager.checkCanceled()
-            return runReadAction { doGetItemIcon(nodeElement, nodeItem) }
-        }
-
-        private fun doGetItemIcon(nodeElement: PsiElement?, nodeItem: Any?): Icon? {
-            return when (nodeElement) {
-                is ParadoxScriptProperty -> {
-                    when {
-                        nodeItem is CwtProperty -> PlsIcons.Nodes.Type
-                        nodeItem is ParadoxScriptProperty -> {
-                            val definitionInfo = nodeItem.definitionInfo
-                            if (definitionInfo != null) return null
-                            PlsIcons.Nodes.Property
-                        }
-                        else -> null
-                    }
-                }
+            return when (nodeItem) {
+                is Items.Type -> PlsIcons.Nodes.Type
+                is Items.Property -> PlsIcons.Nodes.Property
                 else -> null
             }
         }
@@ -240,20 +197,25 @@ abstract class ParadoxEventTreeDiagramProvider(gameType: ParadoxGameType) : Para
         file: VirtualFile?, //umlFile
         provider: ParadoxDefinitionDiagramProvider
     ) : ParadoxDefinitionDiagramProvider.DataModel(project, file, provider) {
-        override fun getModificationTracker(): FilePathBasedModificationTracker {
-            return ParadoxModificationTrackers.ScriptFileTracker("events/**/*.txt")
+        override fun getModificationTracker(): ModificationTracker {
+            val definitionType = ParadoxDefinitionTypes.Event
+            val configGroup = PlsFacade.getConfigGroup(project, provider.gameType)
+            val typeConfig = configGroup.types.get(definitionType) ?: return super.getModificationTracker()
+            val key = CwtConfigManager.getFilePathPatterns(typeConfig).joinToString(";")
+            return ParadoxModificationTrackers.ScriptFileTracker(key)
         }
 
         override fun updateDataModel() {
             //群星原版事件有5000+
-            provider as ParadoxDefinitionDiagramProvider
 
-            val events0 = getDefinitions(ParadoxDefinitionTypes.Event)
+            val definitionType = ParadoxDefinitionTypes.Event
+            val events0 = getDefinitions(definitionType)
             if (events0.isEmpty()) return
             val settings = provider.getDiagramSettings(project)?.state
             val events = events0.filter { settings == null || showNode(it, settings) }
             if (events.isEmpty()) return
 
+            provider as ParadoxDefinitionDiagramProvider
             val nodeMap = mutableMapOf<ParadoxScriptDefinitionElement, Node>()
             val eventMap = mutableMapOf<String, ParadoxScriptDefinitionElement>()
             for (event in events) {
@@ -273,7 +235,7 @@ abstract class ParadoxEventTreeDiagramProvider(gameType: ParadoxGameType) : Para
                     ProgressManager.checkCanceled()
                     val source = nodeMap.get(event) ?: continue
                     val target = eventMap.get(invocation)?.let { nodeMap.get(it) } ?: continue
-                    val edge = Edge(source, target, REL_INVOKE)
+                    val edge = Edge(source, target, Relations.Invoke)
                     edges.add(edge)
                 }
             }