+++ src/main/kotlin/icu/windea/pls/lang/index/ParadoxDefinitionNameIndex.kt
@@ -22,15 +22,7 @@ class ParadoxDefinitionNameIndex : StringStubIndexExtension<ParadoxScriptDefinit
      * 用于快速索引文本格式的名字（其定义类型为`text_format`）。它们是忽略大小写的。
      */
     class TextFormatIndex : StringStubIndexExtension<ParadoxScriptDefinitionElement>() {
-        companion object {
-            private const val VERSION = 71 //2.0.1-dev
-            private const val CACHE_SIZE = 2 * 1024
-        }
-
         override fun getKey() = ParadoxIndexManager.DefinitionNameForTextFormatKey
-
         override fun getVersion() = VERSION
-
-        override fun getCacheSize() = CACHE_SIZE
     }
 }