+++ src/main/kotlin/icu/windea/pls/lang/index/ParadoxIndexManager.kt
@@ -19,11 +19,15 @@ import icu.windea.pls.script.psi.*
 object ParadoxIndexManager {
     val ScriptedVariableNameKey = StubIndexKey.createIndexKey<String, ParadoxScriptScriptedVariable>("paradox.scriptedVariable.name.index")
     val DefinitionNameKey = StubIndexKey.createIndexKey<String, ParadoxScriptDefinitionElement>("paradox.definition.name.index")
-    val DefinitionNameForTextFormatKey = StubIndexKey.createIndexKey<String, ParadoxScriptDefinitionElement>("paradox.definition.name.index.textFormat")
     val DefinitionTypeKey = StubIndexKey.createIndexKey<String, ParadoxScriptDefinitionElement>("paradox.definition.type.index")
     val LocalisationNameKey = StubIndexKey.createIndexKey<String, ParadoxLocalisationProperty>("paradox.localisation.name.index")
-    val LocalisationNameForModifierKey = StubIndexKey.createIndexKey<String, ParadoxLocalisationProperty>("paradox.localisation.name.index.modifier")
     val SyncedLocalisationNameKey = StubIndexKey.createIndexKey<String, ParadoxLocalisationProperty>("paradox.syncedLocalisation.name.index")
+
+    val DefinitionNameForTextFormatKey = StubIndexKey.createIndexKey<String, ParadoxScriptDefinitionElement>("paradox.definition.name.index.textFormat")
+
+    val LocalisationNameForModifierKey = StubIndexKey.createIndexKey<String, ParadoxLocalisationProperty>("paradox.localisation.name.index.modifier")
+    val LocalisationNameForEventKey = StubIndexKey.createIndexKey<String, ParadoxLocalisationProperty>("paradox.localisation.name.index.event")
+    val LocalisationNameForTechKey = StubIndexKey.createIndexKey<String, ParadoxLocalisationProperty>("paradox.localisation.name.index.tech")
 
     val Define by lazy { findFileBasedIndex<ParadoxDefineIndex>() }
     val InlineScriptUsage by lazy { findFileBasedIndex<ParadoxInlineScriptUsageIndex>() }