+++ src/main/kotlin/icu/windea/pls/model/constraints/ParadoxIndexConstraint.kt
@@ -10,20 +10,38 @@ import icu.windea.pls.script.psi.*
 interface ParadoxIndexConstraint<T : PsiElement> {
     val indexKey: StubIndexKey<String, T>
     val ignoreCase: Boolean
+    val inferred: Boolean
 
     enum class Definition(
         override val indexKey: StubIndexKey<String, ParadoxScriptDefinitionElement>,
-        override val ignoreCase: Boolean,
-        val predicate: (definitionType: String) -> Boolean
+        override val ignoreCase: Boolean = false,
+        override val inferred: Boolean = false,
     ) : ParadoxIndexConstraint<ParadoxScriptDefinitionElement> {
-        TextFormat(ParadoxIndexManager.DefinitionNameForTextFormatKey, true, { it == ParadoxDefinitionTypes.TextFormat });
+        TextFormat(ParadoxIndexManager.DefinitionNameForTextFormatKey, ignoreCase = true) {
+            override fun supports(definitionType: String) = definitionType == ParadoxDefinitionTypes.TextFormat
+        },
+        ;
+
+        abstract fun supports(definitionType: String): Boolean
     }
 
     enum class Localisation(
         override val indexKey: StubIndexKey<String, ParadoxLocalisationProperty>,
-        override val ignoreCase: Boolean,
-        val predicate: (name: String) -> Boolean
+        override val ignoreCase: Boolean = false,
+        override val inferred: Boolean = false,
     ) : ParadoxIndexConstraint<ParadoxLocalisationProperty> {
-        Modifier(ParadoxIndexManager.LocalisationNameForModifierKey, true, { it.startsWith("mod_", true) });
+        Modifier(ParadoxIndexManager.LocalisationNameForModifierKey, ignoreCase = true) {
+            override fun supports(name: String) = name.startsWith("mod_", true)
+        },
+        Event(ParadoxIndexManager.LocalisationNameForEventKey, inferred = true) {
+            private val regex = """[\w.]+\.\d+(\.[\w.]*)?""".toRegex()
+            override fun supports(name: String) = name.contains('.') && name.matches(regex)
+        },
+        Tech(ParadoxIndexManager.LocalisationNameForTechKey) {
+            override fun supports(name: String) = name.startsWith("tech_")
+        },
+        ;
+
+        abstract fun supports(name: String): Boolean
     }
 }